+++ src/main/java/com/lxp/aplus/payment/application/usecase/PaymentCommandUseCase.java
@@ -2,10 +2,10 @@ package com.lxp.aplus.payment.application.usecase;
 
 import com.lxp.aplus.common.error.BusinessException;
 import com.lxp.aplus.common.error.code.PaymentErrorCode;
-import com.lxp.aplus.order.application.result.OrderCreateResult;
-import com.lxp.aplus.order.application.usecase.OrderCommandUseCase;
+import com.lxp.aplus.payment.application.result.OrderCreateResult;
 import com.lxp.aplus.payment.application.command.PaymentConfirmCommand;
 import com.lxp.aplus.payment.application.command.PaymentPrepareCommand;
+import com.lxp.aplus.payment.application.port.out.OrderCommandPort;
 import com.lxp.aplus.payment.domain.Payment;
 import com.lxp.aplus.payment.domain.PaymentRepository;
 import com.lxp.aplus.payment.presentation.response.PaymentPrepareResponse;
@@ -19,12 +19,12 @@ import org.springframework.transaction.annotation.Transactional;
 public class PaymentCommandUseCase {
 
     private final PaymentRepository paymentRepository;
-    private final OrderCommandUseCase orderCommandUseCase; // FIXME: Order BC 침범 (의도됨) ⚠️
+    private final OrderCommandPort orderCommandPort;
 
     public PaymentPrepareResponse prepare(PaymentPrepareCommand command) {
 
         // 1. Order 생성 (From Order BC)
-        OrderCreateResult orderResult = orderCommandUseCase.createOrderFromCourseIds(
+        OrderCreateResult orderResult = orderCommandPort.createOrderFromCourseIds(
                 command.userId(),
                 command.courseIds()
         );
@@ -53,7 +53,7 @@ public class PaymentCommandUseCase {
         payment.approve(command.paymentKey(), command.amount());
 
         // 4. Order BC에 완료 통보 (결제 성공 후 Order 상태를 COMPLETED로 변경 요청)
-        orderCommandUseCase.completeOrder(command.orderId(), payment.getPaymentId(), command.amount());
+        orderCommandPort.completeOrder(command.orderId(), payment.getPaymentId(), command.amount());
 
         // 5. Payment 저장
         paymentRepository.save(payment);
