+++ src/main/java/com/lxp/aplus/payment/infrastructure/adapter/OrderCommandAdapter.java
@@ -1,4 +1,4 @@
-package com.lxp.aplus.order.application.usecase;
+package com.lxp.aplus.payment.infrastructure.adapter;
 
 import com.lxp.aplus.common.error.BusinessException;
 import com.lxp.aplus.common.error.code.CourseErrorCode;
@@ -6,22 +6,23 @@ import com.lxp.aplus.common.error.code.OrderErrorCode;
 import com.lxp.aplus.course.domain.Course;
 import com.lxp.aplus.course.domain.CourseRepository;
 import com.lxp.aplus.order.application.CoursePrice;
-import com.lxp.aplus.order.application.result.OrderCreateResult;
+import com.lxp.aplus.payment.application.result.OrderCreateResult;
 import com.lxp.aplus.order.domain.Order;
 import com.lxp.aplus.order.domain.OrderLine;
 import com.lxp.aplus.order.domain.OrderRepository;
+import com.lxp.aplus.payment.application.port.out.OrderCommandPort;
 import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 
-@Service
+@Component
 @RequiredArgsConstructor
 @Transactional
-public class OrderCommandUseCase {
+public class OrderCommandAdapter implements OrderCommandPort {
 
     private final OrderRepository orderRepository;
     private final CourseRepository courseRepository; // FIXME: Course BC 침범 (의도됨) ⚠️
